Fixed up minor issues and added favicon

+++ viewer/static/js/models/piece.js
@@ -8,7 +8,7 @@ define([
     var Piece = Backbone.Model.extend({
         urlRoot: Server.Constants.PIECE_API,
         isHorizontal: function (){
-            return parseInt(this.get('width')) >= parseInt(this.get('height'));
+            return parseFloat(this.get('width')) >= parseFloat(this.get('height'));
         }
     });
